refactor[#345] : 마킹 리스트 api 요청 훅을 features에서 entities로 이동
