+++ src/entities/marking/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./getAddressFromLatLng";
+export * from "./getMarkingList";
