+++ src/entities/marking/constants/index.ts
@@ -1,4 +1,5 @@
 import { API_BASE_URL } from "@/shared/constants";
+import { GetMarkingListRequest } from "../api";
 
 export const REVERSE_GEOCODING_END_POINT = ({
   lat,
@@ -7,3 +8,11 @@ export const REVERSE_GEOCODING_END_POINT = ({
   lat: number;
   lng: number;
 }) => `${API_BASE_URL}/maps/reverse-geocode?lat=${lat}&lng=${lng}`;
+
+export const SEARCH_MARKING_END_POINT = ({
+  southWestLat,
+  southWestLng,
+  northEastLat,
+  northEastLng,
+}: Omit<GetMarkingListRequest, "token">) =>
+  `${API_BASE_URL}/markings/search?southBottomLat=${southWestLat}&northTopLat=${northEastLat}&southLeftLng=${southWestLng}&northRightLng=${northEastLng}`;
