+++ src/features/map/ui/mapMarkers.tsx
@@ -1,6 +1,7 @@
-import { useGetMarkingList } from "@/features/marking/api";
 import { User, Pin, MultiplePin, Cluster } from "@/entities/map/ui";
+import { useGetMarkingList } from "@/entities/marking/api";
 import { API_BASE_URL } from "@/shared/constants";
+import { useResearchMarkingList } from "../hooks";
 import { useMapStore } from "../store";
 
 /*---------- default mode 일 때에만 사용되는 마커입니다. ---------- */
@@ -17,9 +18,15 @@ export const UserMarker = () => {
 };
 
 export const PinMarker = () => {
-  const { data: markersInfo } = useGetMarkingList();
+  const { bounds } = useResearchMarkingList();
+  const { data: markingList } = useGetMarkingList({
+    southWestLat: bounds?.southWest.lat,
+    southWestLng: bounds?.southWest.lng,
+    northEastLat: bounds?.northEast.lat,
+    northEastLng: bounds?.northEast.lng,
+  });
 
-  return markersInfo?.map(({ markingId, lat, lng, images }) => (
+  return markingList?.map(({ markingId, lat, lng, images }) => (
     <Pin
       key={markingId}
       position={{
