+++ src/features/marking/constants/endPoint.ts
@@ -1,14 +1,6 @@
 import { API_BASE_URL } from "@/shared/constants";
-import { GetMarkingListRequest } from "../api";
 
 export const MARKING_END_POINT = {
-  SEARCH_MARKING: ({
-    southWestLat,
-    southWestLng,
-    northEastLat,
-    northEastLng,
-  }: Omit<GetMarkingListRequest, "token">) =>
-    `${API_BASE_URL}/markings/search?southBottomLat=${southWestLat}&northTopLat=${northEastLat}&southLeftLng=${southWestLng}&northRightLng=${northEastLng}`,
   ADD: `${API_BASE_URL}/markings`,
   SAVE_TEMP: `${API_BASE_URL}/markings/temp`,
   DELETE: "/markings",
